[preload-it-custom] Change to directly return the response instead of a blob wrapped in an URL
The caller can always recretae a blob and a URL.createObjectURL() anyway if needed.
But at least, we avoid allocating more (with a potential leak), and we do not need to have an async call to get the response back from the ObjectURL.
When loading a JSON file, this is more effecient and synchronous.