Remove dist from .gitignore
The owner of the package has two versions of the package. One for decelopping, and one exposed as a distribution.
You get the distribution version under the /dist/ folder. But when we do a custom version we have only one version of the package.
The script does not have a 'prepare' script (ran on installation) that would just do 'npm run build' whoch does the dist packagaint/minification.
Without the 'dist' folder, the package is actually not recognized by React.

The minimum change here is to not ignore the 'dist' folder, this exposes the 'dist' folder that we generated (again with 'npm run build').
When we install proload-it-custom, we get the source files and the dist version installed, but it not an issue as it is an internal usage anyway.

It is not worth to do exactly the same process as the package owner (ignore dist, unignore sources on one side, unignore dist, ignore sources on the other side).
We want automation to be able to easily pick the stuff up with minimum local iterations.

This commit also contains the generated dist files.